feat: Consolidate repo
