--- disabled/test-worker/CHANGELOG.md
@@ -1,9 +1,0 @@
-# test-worker
-
-## null
-
-### Patch Changes
-
--   Updated dependencies [d0929c4]
-    -   @crossbuild/functions@0.3.0
-    -   @crossbuild/core@0.1.0
