--- disabled/test-worker/package.json
@@ -1,21 +1,0 @@
-{
-	"name": "test-worker",
-	"private": true,
-	"main": "src/index.ts",
-	"type": "module",
-	"scripts": {
-		"clean": "rimraf dist",
-		"deploy": "wrangler publish",
-		"lint": "eslint . --fix",
-		"tail": "wrangler tail",
-		"dev": "dotenv -- wrangler dev"
-	},
-	"dependencies": {
-		"@cloudflare/workers-types": "^4.20230807.0",
-		"@crossbuild/core": "workspace:^",
-		"@crossbuild/functions": "workspace:^",
-		"@crossbuild/worker": "workspace:^",
-		"wrangler": "^3.5.0"
-	},
-	"version": null
-}
