--- disabled/test-worker/src/components/commands/sunday.ts
@@ -1,15 +1,0 @@
-import { Component, CrossBuild, ReceivedInteraction } from "@crossbuild/core"
-import { todayIsSunday } from "../../customChecks"
-
-export default class Cmd extends Component {
-    constructor(client: CrossBuild) {
-        super("ping", "command", client, {
-            description: "Ping!",
-            customChecks: [todayIsSunday]
-        })
-    }
-
-    public override async run(interaction: ReceivedInteraction) {
-        interaction.reply("It's sunday!")
-    }
-}
