--- disabled/test-worker/src/components/selectMenus/ping.ts
@@ -1,14 +1,0 @@
-import { Component, CrossBuild, ReceivedInteraction } from "@crossbuild/core"
-
-export default class Cmd extends Component {
-    constructor(client: CrossBuild) {
-        super("selectping", "selectMenu", client, {
-        })
-    }
-
-    public override async run(interaction: ReceivedInteraction) {
-        await interaction.reply({
-            content: `Dropdown pong: ${interaction.selectMenuValues?.join(", ")}`
-        })
-    }
-}
