--- disabled/test-worker/tsconfig.json
@@ -1,8 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"include": ["src/**/*.ts", "src/*.ts"],
-	"compilerOptions": {
-		"outDir": "./dist",
-		"rootDir": "."
-	}
-}
