--- disabled/worker/README.md
@@ -1,8 +1,0 @@
-<div align="center">
-<img width="150" height="150" src="https://cdn.buape.com/crossbuild.png" alt="CrossBuild Logo" />
-
-<h1 align="center"><b>@crossbuild/worker</b></h1>
-
-This package contains the Cloudflare Worker HTTP Discord module for [CrossBuild](https://npmjs.com/package/crossbuild).
-
-</div>
