--- disabled/worker/package.json
@@ -1,51 +1,0 @@
-{
-	"name": "@crossbuild/worker",
-	"description": "Discord modules for HTTP Cloudflare Workers for CrossBuild",
-	"type": "module",
-	"version": "0.0.1",
-	"private": true,
-	"authors": [
-		"Shadow <shadow@buape.com>",
-		"Codeize <codeize@buape.com>"
-	],
-	"scripts": {
-		"lint": "eslint . --fix",
-		"build": "unbuild",
-		"clean": "rimraf dist"
-	},
-	"main": "./dist/index.cjs",
-	"module": "./dist/index.mjs",
-	"types": "./dist/index.d.ts",
-	"exports": {
-		".": {
-			"types": "./dist/index.d.ts",
-			"import": "./dist/index.mjs",
-			"require": "./dist/index.cjs"
-		}
-	},
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/buape/crossbuild.git",
-		"directory": "packages/guilded"
-	},
-	"license": "MIT",
-	"bugs": {
-		"url": "https://github.com/buape/crossbuild/issues"
-	},
-	"files": [
-		"dist",
-		"LICENSE"
-	],
-	"engines": {
-		"node": ">=18.17.0"
-	},
-	"dependencies": {
-		"@cloudflare/workers-types": "4.20230807.0",
-		"@crossbuild/core": "workspace:^",
-		"@crossbuild/types": "workspace:^",
-		"@discordjs/rest": "^2.0.0",
-		"discord-api-types": "^0.37.50",
-		"discord-verify": "^1.2.0",
-		"itty-router": "^4.0.14"
-	}
-}
