--- disabled/worker/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./classes/WorkerModule"
-export * from "./classes/WorkerReceivedInteraction"
