--- disabled/worker/typedoc.json
@@ -1,4 +1,0 @@
-{
-	"extends": ["../../typedoc.base.json"],
-	"entryPoints": ["src/index.ts"]
-}
