+++ packages/core/package.json
@@ -39,11 +39,12 @@
 		"node": ">=18.17.0"
 	},
 	"dependencies": {
-		"@crossbuild/functions": "workspace:^",
-		"@crossbuild/types": "workspace:^",
-		"@discordjs/collection": "^1.5.3"
+		"@discordjs/collection": "^1.5.3",
+		"discord-api-types": "^0.37.54",
+		"node-fetch": "2"
 	},
 	"devDependencies": {
-		"@types/node": "^18.17.5"
+		"@types/node": "^18.17.5",
+		"@types/node-fetch": "^2.6.4"
 	}
 }
