+++ packages/core/src/functions/generateTimestamp.ts
