+++ packages/core/src/functions/titleCase.ts
