+++ packages/discord/package.json
@@ -40,7 +40,6 @@
 	},
 	"dependencies": {
 		"@crossbuild/core": "workspace:^",
-		"@crossbuild/types": "workspace:^",
 		"discord.js": "^14.13.0"
 	}
 }
