--- packages/functions/CHANGELOG.md
@@ -1,51 +1,0 @@
-# @crossbuild/functions
-
-## 0.3.1
-
-### Patch Changes
-
--   294aba0: chore: Update dependencies
--   8a6405e: chore: update dependencies
--   Updated dependencies [294aba0]
--   Updated dependencies [8a6405e]
-    -   @crossbuild/types@0.2.1
-
-## 0.3.0
-
-### Minor Changes
-
--   d0929c4: Switch to a modular system for managing different platforms
-
-### Patch Changes
-
--   Updated dependencies [d0929c4]
-    -   @crossbuild/types@0.2.0
-
-## 0.2.1
-
-### Patch Changes
-
--   fe65ae6: Add client permissions checks for Components
--   Updated dependencies [35dfe13]
-    -   @crossbuild/types@0.1.1
-
-## 0.2.0
-
-### Minor Changes
-
--   051ed72: Add a new options handler that is more type-safe to use
-
-### Patch Changes
-
--   d9f0aab: Update to discord.js v14.12.1, move interaction data building to the listeners
-
-## 0.1.0
-
-### Minor Changes
-
--   26276e7: Initial release of CrossBuild
-
-### Patch Changes
-
--   Updated dependencies [26276e7]
-    -   @crossbuild/types@0.1.0
