--- packages/functions/README.md
@@ -1,8 +1,0 @@
-<div align="center">
-<img width="150" height="150" src="https://cdn.buape.com/crossbuild.png" alt="CrossBuild Logo" />
-
-<h1 align="center"><b>@crossbuild/functions</b></h1>
-
-This package is contains useful utility functions for use with [CrossBuild](https://npmjs.com/package/crossbuild).
-
-</div>
