--- packages/functions/package.json
@@ -1,50 +1,0 @@
-{
-	"name": "@crossbuild/functions",
-	"description": "Useful utility functions, mainly for use in Discord bots",
-	"type": "module",
-	"version": "0.3.1",
-	"authors": [
-		"Shadow <shadow@buape.com>",
-		"Codeize <codeize@buape.com>"
-	],
-	"scripts": {
-		"lint": "eslint . --fix",
-		"build": "unbuild",
-		"clean": "rimraf dist"
-	},
-	"main": "./dist/index.cjs",
-	"module": "./dist/index.mjs",
-	"types": "./dist/index.d.ts",
-	"exports": {
-		".": {
-			"types": "./dist/index.d.ts",
-			"import": "./dist/index.mjs",
-			"require": "./dist/index.cjs"
-		}
-	},
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/buape/crossbuild.git",
-		"directory": "packages/functions"
-	},
-	"license": "MIT",
-	"bugs": {
-		"url": "https://github.com/buape/crossbuild/issues"
-	},
-	"files": [
-		"dist",
-		"LICENSE"
-	],
-	"engines": {
-		"node": ">=18.17.0"
-	},
-	"dependencies": {
-		"@crossbuild/types": "workspace:^",
-		"discord-api-types": "^0.37.54",
-		"node-fetch": "^2.6.12"
-	},
-	"devDependencies": {
-		"@types/node": "^18.17.5",
-		"@types/node-fetch": "^2.6.4"
-	}
-}
