--- packages/functions/src/index.ts
@@ -1,5 +1,0 @@
-export * from "./generateEmbed.js"
-export * from "./generateTimestamp.js"
-export * from "./getFiles.js"
-export * from "./titleCase.js"
-export * from "./uploadHaste.js"
