--- packages/types/CHANGELOG.md
@@ -1,26 +1,0 @@
-# @crossbuild/types
-
-## 0.2.1
-
-### Patch Changes
-
--   294aba0: chore: Update dependencies
--   8a6405e: chore: update dependencies
-
-## 0.2.0
-
-### Minor Changes
-
--   d0929c4: Switch to a modular system for managing different platforms
-
-## 0.1.1
-
-### Patch Changes
-
--   35dfe13: Add an enum for Guilded emoji IDs
-
-## 0.1.0
-
-### Minor Changes
-
--   26276e7: Initial release of CrossBuild
